minor, provide fixc_msg_add_tag() and fixc_msg_add_fld() to add tags and fields respectively
Adding fields is always by means of the fixc_fld_s struct and the
field type must not be FIXC_TYP_OFF (because the offset would be
different for different messages).  Instead, fixc_msg_add_tag() can
be used in that case, the value supplied will be copied and a msg
specific offset will be used.

Signed-off-by: Sebastian Freundt <freundt@ga-group.nl>